refactor: 기타 리팩토링
