--- tomcat/src/main/java/org/apache/coyote/http11/HttpMethod.java
@@ -1,4 +1,0 @@
-package org.apache.coyote.http11;
-
-public class HttpMethod {
-}
