+++ tomcat/src/main/java/org/apache/coyote/http11/controller/LoginController.java
@@ -37,6 +37,7 @@ public class LoginController extends AbstractController {
 
             final var session = request.getSession();
             session.setAttribute("user", user.get());
+
             response.addCookie(Cookie.ofJSessionId(session.getId()));
             response.setStatusCode(StatusCode.FOUND);
             response.sendRedirect("/index.html");
