--- tomcat/src/main/java/org/apache/coyote/http11/exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package org.apache.coyote.http11.exception;
-
-public class BadRequestException extends RuntimeException {
-
-    public BadRequestException(String message) {
-        super(message);
-    }
-}
