+++ tomcat/src/main/java/org/apache/coyote/http11/model/HttpResponse.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.model;
 
-import org.apache.coyote.http11.controller.StaticResourceHandler;
+import org.apache.coyote.http11.controller.StaticResourceController;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -16,7 +16,7 @@ public class HttpResponse {
         this.version = "HTTP/1.1";
         this.statusCode = StatusCode.INTERNAL_SERVER_ERROR;
         this.headers = new HashMap<>();
-        setHeader("Content-Type", "text/html;charset=utf-8");
+        setHeader("Content-Type", StaticResourceController.DEFAULT_CONTENT_TYPE);
         setHeader("Content-Length", "0");
         this.body = "";
     }
@@ -30,7 +30,9 @@ public class HttpResponse {
         }
         sb.append("\r\n");
 
-        sb.append(body != null ? body : "");
+        if(body != null) {
+            sb.append(body);
+        }
 
         return sb.toString().getBytes();
     }
