+++ tomcat/src/main/java/org/apache/coyote/http11/model/QueryParameter.java
@@ -6,10 +6,6 @@ import java.util.Map;
 
 public class QueryParameter {
 
-    private static final String PAIR_DELIMITER = "&";
-    private static final String KEY_VALUE_DELIMITER = "=";
-    private static final int VALID_PAIR_COUNT = 2;
-
     private final Map<String, String> queryParameter;
 
     public QueryParameter() {
@@ -33,9 +29,9 @@ public class QueryParameter {
         if (data == null || data.isBlank()) {
             return;
         }
-        String[] pairs = data.split(PAIR_DELIMITER);
+        String[] pairs = data.split("&");
         for (String pair : pairs) {
-            String[] keyValue = pair.split(KEY_VALUE_DELIMITER, VALID_PAIR_COUNT);
+            String[] keyValue = pair.split("=", 2);
             if (keyValue[0].isBlank()) {
                 continue;
             }
