+++ tomcat/src/main/java/org/apache/coyote/http11/processor/Http11Processor.java
@@ -1,11 +1,11 @@
 package org.apache.coyote.http11.processor;
 
-import org.apache.coyote.Processor;
 import org.apache.coyote.http11.controller.Controller;
 import org.apache.coyote.http11.controller.ControllerMapper;
 import org.apache.coyote.http11.controller.LoginController;
 import org.apache.coyote.http11.controller.RegisterController;
-import org.apache.coyote.http11.controller.StaticResourceHandler;
+import org.apache.coyote.http11.controller.StaticResourceController;
+import org.apache.coyote.http11.model.HttpRequest;
 import org.apache.coyote.http11.model.HttpResponse;
 import org.apache.coyote.http11.util.HttpRequestUtil;
 import org.apache.coyote.http11.session.SessionManager;
@@ -19,12 +19,12 @@ public class Http11Processor implements Runnable, Processor {
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
 
     private final Socket connection;
-    private final StaticResourceHandler resourceHandler;
+    private final StaticResourceController resourceHandler;
     private final ControllerMapper controllerMapper;
 
     public Http11Processor(final Socket connection) {
         this.connection = connection;
-        this.resourceHandler = new StaticResourceHandler();
+        this.resourceHandler = new StaticResourceController();
         this.controllerMapper = new ControllerMapper();
         initializeControllers();
     }
@@ -53,16 +53,22 @@ public class Http11Processor implements Runnable, Processor {
             httpRequest.setSession(session);
 
             final Controller controller = controllerMapper.getController(httpRequest.getPath());
-            if (controller != null) {
-                controller.service(httpRequest, httpResponse);
-            } else {
-                resourceHandler.execute(httpRequest, httpResponse);
-            }
+
+            executeHandler(controller, httpRequest, httpResponse);
 
             outputStream.write(httpResponse.getBytes());
             outputStream.flush();
         } catch (Exception e) {
             log.error(e.getMessage(), e);
         }
+    }
+
+    private void executeHandler(final Controller controller, final HttpRequest httpRequest,
+                                final HttpResponse httpResponse) throws Exception {
+        if (controller != null) {
+            controller.service(httpRequest, httpResponse);
+            return;
+        }
+        resourceHandler.execute(httpRequest, httpResponse);
     }
 }
