+++ tomcat/src/main/java/org/apache/coyote/http11/processor/Processor.java
@@ -14,7 +14,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.apache.coyote;
+package org.apache.coyote.http11.processor;
 
 import java.net.Socket;
